combine functionality of old + new format parsers
